add package.json, gulp config
remove doc comment from morphic.js
