+++ src/003_retina.js
@@ -1,0 +1,365 @@
+// Retina Display Support //////////////////////////////////////////////
+
+/*
+    By default retina support gets installed when Morphic.js loads. There
+    are two global functions that let you test for retina availability:
+
+        isRetinaSupported() - Boolean, whether retina support is available
+        isRetinaEnabled()   - Boolean, whether currently in retina mode
+
+    and two more functions that let you control retina support if it is
+    available:
+
+        enableRetinaSupport()
+        disableRetinaSupport()
+
+    Both of these internally test whether retina is available, so they are
+    safe to call directly.
+
+    Even when in retina mode it often makes sense to use non-high-resolution
+    canvasses for simple shapes in order to save system resources and
+    optimize performance. Examples are costumes and backgrounds in Snap.
+    In Morphic you can create new canvas elements using
+
+        newCanvas(extentPoint [, nonRetinaFlag])
+
+    If retina support is enabled such new canvasses will automatically be
+    high-resolution canvasses, unless the newCanvas() function is given an
+    otherwise optional second Boolean <true> argument that explicitly makes
+    it a non-retina canvas.
+
+    Not the whole canvas API is supported by Morphic's retina utilities.
+    Especially if your code uses putImageData() you will want to "downgrade"
+    a target high-resolution canvas to a normal-resolution ("non-retina")
+    one before using
+
+        normalizeCanvas(aCanvas [, copyFlag])
+
+    This will change the target canvas' resolution in place (!). If you
+    pass in the optional second Boolean <true> flag the function returns
+    a non-retina copy and leaves the target canvas unchanged. An example
+    of this normalize mechanism is converting the penTrails layer of Snap's
+    stage (high-resolution) into a sprite-costume (normal resolution).
+*/
+
+function enableRetinaSupport() {
+    /*
+        === contributed by Bartosz Leper ===
+
+        This installs a series of utilities that allow using Canvas the same way
+        on retina and non-retina displays. If the display is a retina one, the
+        underlying dimensions of the Canvas elements are doubled, but this will
+        be transparent to the code that uses Canvas. All dimensions read or
+        written to the Canvas element will be scaled appropriately.
+
+        NOTE: This implementation is not exhaustive; it only implements what is
+        needed by the Snap! UI.
+
+        [Jens]: like all other retina screen support implementations I've seen
+        Bartosz's patch also does not address putImageData() compatibility when
+        mixing retina-enabled and non-retina canvasses. If you need to manipulate
+        pixels in such mixed canvasses, make sure to "downgrade" them all using
+        normalizeCanvas() below.
+    */
+
+    // Get the window's pixel ratio for canvas elements.
+    // See: http://www.html5rocks.com/en/tutorials/canvas/hidpi/
+    var ctx = document.createElement("canvas").getContext("2d"),
+        backingStorePixelRatio = ctx.webkitBackingStorePixelRatio ||
+            ctx.mozBackingStorePixelRatio ||
+            ctx.msBackingStorePixelRatio ||
+            ctx.oBackingStorePixelRatio ||
+            ctx.backingStorePixelRatio || 1,
+
+        // Unfortunately, it's really hard to make this work well when changing
+        // zoom level, so let's leave it like this right now, and stick to
+        // whatever the ratio was in the beginning.
+
+        // originalDevicePixelRatio = window.devicePixelRatio,
+
+        // [Jens]: As of summer 2016 non-integer devicePixelRatios lead to
+        // artifacts when blitting images onto canvas elements in all browsers
+        // except Chrome, especially Firefox, Edge, IE (Safari doesn't even
+        // support retina mode as implemented here).
+        // therefore - to ensure crisp fonts - use the ceiling of whatever
+        // the devicePixelRatio is. This needs more memory, but looks nicer.
+
+        originalDevicePixelRatio = Math.ceil(window.devicePixelRatio),
+
+        canvasProto = HTMLCanvasElement.prototype,
+        contextProto = CanvasRenderingContext2D.prototype,
+
+        // [Jens]: keep track of original properties in a dictionary
+        // so they can be iterated over and restored
+        uber = {
+            drawImage: contextProto.drawImage,
+            getImageData: contextProto.getImageData,
+
+            width: Object.getOwnPropertyDescriptor(
+                canvasProto,
+                'width'
+            ),
+            height: Object.getOwnPropertyDescriptor(
+                canvasProto,
+                'height'
+            ),
+            shadowOffsetX: Object.getOwnPropertyDescriptor(
+                contextProto,
+                'shadowOffsetX'
+            ),
+            shadowOffsetY: Object.getOwnPropertyDescriptor(
+                contextProto,
+                'shadowOffsetY'
+            ),
+            shadowBlur: Object.getOwnPropertyDescriptor(
+                contextProto,
+                'shadowBlur'
+            )
+        };
+
+    // [Jens]: only install retina utilities if the display supports them
+    if (backingStorePixelRatio === originalDevicePixelRatio) {return; }
+    // [Jens]: check whether properties can be overridden, needed for Safari
+    if (Object.keys(uber).some(function (any) {
+        var prop = uber[any];
+        return prop.hasOwnProperty('configurable') && (!prop.configurable);
+    })) {return; }
+
+    function getPixelRatio(imageSource) {
+        return imageSource.isRetinaEnabled ?
+            (originalDevicePixelRatio || 1) / backingStorePixelRatio : 1;
+    }
+
+    canvasProto._isRetinaEnabled = true;
+    // [Jens]: remember the original non-retina properties,
+    // so they can be restored again
+    canvasProto._bak = uber;
+
+    Object.defineProperty(canvasProto, 'isRetinaEnabled', {
+        get: function() {
+            return this._isRetinaEnabled;
+        },
+        set: function(enabled) {
+            var prevPixelRatio = getPixelRatio(this),
+                prevWidth = this.width,
+                prevHeight = this.height;
+
+            this._isRetinaEnabled = enabled;
+            if (getPixelRatio(this) != prevPixelRatio) {
+                this.width = prevWidth;
+                this.height = prevHeight;
+            }
+        },
+        configurable: true // [Jens]: allow to be deleted an reconfigured
+    });
+
+    Object.defineProperty(canvasProto, 'width', {
+        get: function() {
+            return uber.width.get.call(this) / getPixelRatio(this);
+        },
+        set: function(width) {
+            try { // workaround one of FF's dreaded NS_ERROR_FAILURE bugs
+                // this should be taken out as soon as FF gets fixed again
+                var pixelRatio = getPixelRatio(this),
+                    context;
+                uber.width.set.call(this, width * pixelRatio);
+                context = this.getContext('2d');
+                context.restore();
+                context.save();
+                context.scale(pixelRatio, pixelRatio);
+            } catch (err) {
+                console.log('Retina Display Support Problem', err);
+                uber.width.set.call(this, width);
+            }
+        }
+    });
+
+    Object.defineProperty(canvasProto, 'height', {
+        get: function() {
+            return uber.height.get.call(this) / getPixelRatio(this);
+        },
+        set: function(height) {
+            var pixelRatio = getPixelRatio(this),
+                context;
+            uber.height.set.call(this, height * pixelRatio);
+            context = this.getContext('2d');
+            context.restore();
+            context.save();
+            context.scale(pixelRatio, pixelRatio);
+        }
+    });
+
+    contextProto.drawImage = function(image) {
+        var pixelRatio = getPixelRatio(image),
+            sx, sy, sWidth, sHeight, dx, dy, dWidth, dHeight;
+
+        // Different signatures of drawImage() method have different
+        // parameter assignments.
+        switch (arguments.length) {
+            case 9:
+                sx = arguments[1];
+                sy = arguments[2];
+                sWidth = arguments[3];
+                sHeight = arguments[4];
+                dx = arguments[5];
+                dy = arguments[6];
+                dWidth = arguments[7];
+                dHeight = arguments[8];
+                break;
+
+            case 5:
+                sx = 0;
+                sy = 0;
+                sWidth = image.width;
+                sHeight = image.height;
+                dx = arguments[1];
+                dy = arguments[2];
+                dWidth = arguments[3];
+                dHeight = arguments[4];
+                break;
+
+            case 3:
+                sx = 0;
+                sy = 0;
+                sWidth = image.width;
+                sHeight = image.height;
+                dx = arguments[1];
+                dy = arguments[2];
+                dWidth = image.width;
+                dHeight = image.height;
+                break;
+
+            default:
+                throw Error('Called drawImage() with ' + arguments.length +
+                    ' arguments');
+        }
+        uber.drawImage.call(
+            this, image,
+            sx * pixelRatio, sy * pixelRatio,
+            sWidth * pixelRatio, sHeight * pixelRatio,
+            dx, dy,
+            dWidth, dHeight);
+    };
+
+    contextProto.getImageData = function(sx, sy, sw, sh) {
+        var pixelRatio = getPixelRatio(this.canvas);
+        return uber.getImageData.call(
+            this,
+            sx * pixelRatio, sy * pixelRatio,
+            sw * pixelRatio, sh * pixelRatio);
+    };
+
+    Object.defineProperty(contextProto, 'shadowOffsetX', {
+        get: function() {
+            return uber.shadowOffsetX.get.call(this) /
+                getPixelRatio(this.canvas);
+        },
+        set: function(offset) {
+            var pixelRatio = getPixelRatio(this.canvas);
+            uber.shadowOffsetX.set.call(this, offset * pixelRatio);
+        }
+    });
+
+    Object.defineProperty(contextProto, 'shadowOffsetY', {
+        get: function() {
+            return uber.shadowOffsetY.get.call(this) /
+                getPixelRatio(this.canvas);
+        },
+        set: function(offset) {
+            var pixelRatio = getPixelRatio(this.canvas);
+            uber.shadowOffsetY.set.call(this, offset * pixelRatio);
+        }
+    });
+
+    Object.defineProperty(contextProto, 'shadowBlur', {
+        get: function() {
+            return uber.shadowBlur.get.call(this) /
+                getPixelRatio(this.canvas);
+        },
+        set: function(blur) {
+            var pixelRatio = getPixelRatio(this.canvas);
+            uber.shadowBlur.set.call(this, blur * pixelRatio);
+        }
+    });
+}
+
+function isRetinaSupported () {
+    var ctx = document.createElement("canvas").getContext("2d"),
+        backingStorePixelRatio = ctx.webkitBackingStorePixelRatio ||
+            ctx.mozBackingStorePixelRatio ||
+            ctx.msBackingStorePixelRatio ||
+            ctx.oBackingStorePixelRatio ||
+            ctx.backingStorePixelRatio || 1,
+        canvasProto = HTMLCanvasElement.prototype,
+        contextProto = CanvasRenderingContext2D.prototype,
+        uber = {
+            drawImage: contextProto.drawImage,
+            getImageData: contextProto.getImageData,
+
+            width: Object.getOwnPropertyDescriptor(
+                canvasProto,
+                'width'
+            ),
+            height: Object.getOwnPropertyDescriptor(
+                canvasProto,
+                'height'
+            ),
+            shadowOffsetX: Object.getOwnPropertyDescriptor(
+                contextProto,
+                'shadowOffsetX'
+            ),
+            shadowOffsetY: Object.getOwnPropertyDescriptor(
+                contextProto,
+                'shadowOffsetY'
+            ),
+            shadowBlur: Object.getOwnPropertyDescriptor(
+                contextProto,
+                'shadowBlur'
+            )
+        };
+    return backingStorePixelRatio !== window.devicePixelRatio &&
+        !(Object.keys(uber).some(function (any) {
+                var prop = uber[any];
+                return prop.hasOwnProperty('configurable') && (!prop.configurable);
+            })
+        );
+}
+
+function isRetinaEnabled () {
+    return HTMLCanvasElement.prototype.hasOwnProperty('_isRetinaEnabled');
+}
+
+function disableRetinaSupport() {
+    // uninstalls Retina utilities. Make sure to re-create every Canvas
+    // element afterwards
+    var canvasProto, contextProto, uber;
+    if (!isRetinaEnabled()) {return; }
+    canvasProto = HTMLCanvasElement.prototype;
+    contextProto = CanvasRenderingContext2D.prototype;
+    uber = canvasProto._bak;
+    Object.defineProperty(canvasProto, 'width', uber.width);
+    Object.defineProperty(canvasProto, 'height', uber.height);
+    contextProto.drawImage = uber.drawImage;
+    contextProto.getImageData = uber.getImageData;
+    Object.defineProperty(contextProto, 'shadowOffsetX', uber.shadowOffsetX);
+    Object.defineProperty(contextProto, 'shadowOffsetY', uber.shadowOffsetY);
+    Object.defineProperty(contextProto, 'shadowBlur', uber.shadowBlur);
+    delete canvasProto._isRetinaEnabled;
+    delete canvasProto.isRetinaEnabled;
+    delete canvasProto._bak;
+}
+
+function normalizeCanvas(aCanvas, getCopy) {
+    // make sure aCanvas is non-retina, otherwise convert it in place (!)
+    // or answer a normalized copy if the "getCopy" flag is <true>
+    var cpy;
+    if (!aCanvas.isRetinaEnabled) {return aCanvas; }
+    cpy = newCanvas(new Point(aCanvas.width, aCanvas.height), true);
+    cpy.getContext('2d').drawImage(aCanvas, 0, 0);
+    if (getCopy) {return cpy; }
+    aCanvas.isRetinaEnabled = false;
+    aCanvas.width = cpy.width;
+    aCanvas.height = cpy.height;
+    aCanvas.getContext('2d').drawImage(cpy, 0, 0);
+    return aCanvas;
+}
