+++ src/morphs/012_colorpalette.js
@@ -1,0 +1,115 @@
+var ColorPaletteMorph;
+
+// ColorPaletteMorph inherits from Morph:
+
+ColorPaletteMorph.prototype = new Morph();
+ColorPaletteMorph.prototype.constructor = ColorPaletteMorph;
+ColorPaletteMorph.uber = Morph.prototype;
+
+// ColorPaletteMorph instance creation:
+
+function ColorPaletteMorph(target, sizePoint) {
+    this.init(
+        target || null,
+        sizePoint || new Point(80, 50)
+    );
+}
+
+ColorPaletteMorph.prototype.init = function (target, size) {
+    ColorPaletteMorph.uber.init.call(this);
+    this.target = target;
+    this.targetSetter = 'color';
+    this.silentSetExtent(size);
+    this.choice = null;
+    this.drawNew();
+};
+
+ColorPaletteMorph.prototype.drawNew = function () {
+    var context, ext, x, y, h, l;
+
+    ext = this.extent();
+    this.image = newCanvas(this.extent());
+    context = this.image.getContext('2d');
+    this.choice = new Color();
+    for (x = 0; x <= ext.x; x += 1) {
+        h = 360 * x / ext.x;
+        for (y = 0; y <= ext.y; y += 1) {
+            l = 100 - (y / ext.y * 100);
+            context.fillStyle = 'hsl(' + h + ',100%,' + l + '%)';
+            context.fillRect(x, y, 1, 1);
+        }
+    }
+};
+
+ColorPaletteMorph.prototype.mouseMove = function (pos) {
+    this.choice = this.getPixelColor(pos);
+    this.updateTarget();
+};
+
+ColorPaletteMorph.prototype.mouseDownLeft = function (pos) {
+    this.choice = this.getPixelColor(pos);
+    this.updateTarget();
+};
+
+ColorPaletteMorph.prototype.updateTarget = function () {
+    if (this.target instanceof Morph && this.choice !== null) {
+        if (this.target[this.targetSetter] instanceof Function) {
+            this.target[this.targetSetter](this.choice);
+        } else {
+            this.target[this.targetSetter] = this.choice;
+            this.target.drawNew();
+            this.target.changed();
+        }
+    }
+};
+
+// ColorPaletteMorph menu:
+
+ColorPaletteMorph.prototype.developersMenu = function () {
+    var menu = ColorPaletteMorph.uber.developersMenu.call(this);
+    menu.addLine();
+    menu.addItem(
+        'set target',
+        "setTarget",
+        'choose another morph.jsk\nwhose color property\n will be' +
+        ' controlled by this one'
+    );
+    return menu;
+};
+
+ColorPaletteMorph.prototype.setTarget = function () {
+    var choices = this.overlappedMorphs(),
+        menu = new MenuMorph(this, 'choose target:'),
+        myself = this;
+
+    choices.push(this.world());
+    choices.forEach(function (each) {
+        menu.addItem(each.toString().slice(0, 50), function () {
+            myself.target = each;
+            myself.setTargetSetter();
+        });
+    });
+    if (choices.length === 1) {
+        this.target = choices[0];
+        this.setTargetSetter();
+    } else if (choices.length > 0) {
+        menu.popUpAtHand(this.world());
+    }
+};
+
+ColorPaletteMorph.prototype.setTargetSetter = function () {
+    var choices = this.target.colorSetters(),
+        menu = new MenuMorph(this, 'choose target property:'),
+        myself = this;
+
+    choices.forEach(function (each) {
+        menu.addItem(each, function () {
+            myself.targetSetter = each;
+        });
+    });
+    if (choices.length === 1) {
+        this.targetSetter = choices[0];
+    } else if (choices.length > 0) {
+        menu.popUpAtHand(this.world());
+    }
+};
