+++ src/morphs/014_colorpicker.js
@@ -1,0 +1,64 @@
+
+
+// ColorPickerMorph inherits from Morph:
+
+ColorPickerMorph.prototype = new Morph();
+ColorPickerMorph.prototype.constructor = ColorPickerMorph;
+ColorPickerMorph.uber = Morph.prototype;
+
+// ColorPickerMorph instance creation:
+
+function ColorPickerMorph(defaultColor) {
+    this.init(defaultColor || new Color(255, 255, 255));
+}
+
+ColorPickerMorph.prototype.init = function (defaultColor) {
+    this.choice = defaultColor;
+    ColorPickerMorph.uber.init.call(this);
+    this.color = new Color(255, 255, 255);
+    this.silentSetExtent(new Point(80, 80));
+    this.drawNew();
+};
+
+ColorPickerMorph.prototype.drawNew = function () {
+    ColorPickerMorph.uber.drawNew.call(this);
+    this.buildSubmorphs();
+};
+
+ColorPickerMorph.prototype.buildSubmorphs = function () {
+    var cpal, gpal, x, y;
+
+    this.children.forEach(function (child) {
+        child.destroy();
+    });
+    this.children = [];
+    this.feedback = new Morph();
+    this.feedback.color = this.choice;
+    this.feedback.setExtent(new Point(20, 20));
+    cpal = new ColorPaletteMorph(
+        this.feedback,
+        new Point(this.width(), 50)
+    );
+    gpal = new GrayPaletteMorph(
+        this.feedback,
+        new Point(this.width(), 5)
+    );
+    cpal.setPosition(this.bounds.origin);
+    this.add(cpal);
+    gpal.setPosition(cpal.bottomLeft());
+    this.add(gpal);
+    x = (gpal.left() +
+        Math.floor((gpal.width() - this.feedback.width()) / 2));
+    y = gpal.bottom() + Math.floor((this.bottom() -
+        gpal.bottom() - this.feedback.height()) / 2);
+    this.feedback.setPosition(new Point(x, y));
+    this.add(this.feedback);
+};
+
+ColorPickerMorph.prototype.getChoice = function () {
+    return this.feedback.color;
+};
+
+ColorPickerMorph.prototype.rootForGrab = function () {
+    return this;
+};
