+++ src/morphs/015_blinker.js
@@ -1,0 +1,29 @@
+
+// can be used for text cursors
+
+var BlinkerMorph;
+
+// BlinkerMorph inherits from Morph:
+
+BlinkerMorph.prototype = new Morph();
+BlinkerMorph.prototype.constructor = BlinkerMorph;
+BlinkerMorph.uber = Morph.prototype;
+
+// BlinkerMorph instance creation:
+
+function BlinkerMorph(rate) {
+    this.init(rate);
+}
+
+BlinkerMorph.prototype.init = function (rate) {
+    BlinkerMorph.uber.init.call(this);
+    this.color = new Color(0, 0, 0);
+    this.fps = rate || 2;
+    this.drawNew();
+};
+
+// BlinkerMorph stepping:
+
+BlinkerMorph.prototype.step = function () {
+    this.toggleVisibility();
+};
